minor updates and comments

+++ tisdon/src/components/TaskList/TaskDate.js
@@ -3,7 +3,7 @@ import './TaskDate.css';
 
 const TaskDate = (props) => {
   const day = props.date.toLocaleString('en-US', { day: '2-digit' });
-  const month = props.date.toLocaleString('en-US', { month: 'long' });
+  const month = props.date.toLocaleString('en-US', { month: 'short' });
   const year = props.date.getFullYear();
 
   return (
